lab12/requirements.txt
azure-search-documents
python-dotenv
openai